Hoisted utility find_parameter_by_id from the parameters REST handlers to the top level mod.rs in REST.
